fix: resolve Windows build failure with platform-specific privilege code
- Split privilege.go into platform-specific files using build tags
- Created privilege_unix.go for Unix-specific syscalls (Setrlimit, Credential, etc.)
- Created privilege_windows.go for Windows-compatible implementations
- Separated platform-specific commands and environment variables
- All platforms now compile successfully
