+++ pkg/security/privilege.go
@@ -1,13 +1,8 @@
 package security
 
 import (
-	"fmt"
 	"os"
-	"os/exec"
-	"os/user"
 	"runtime"
-	"strconv"
-	"syscall"
 )
 
 // PrivilegeManager handles privilege separation and dropping
@@ -29,59 +24,7 @@ func NewPrivilegeManager() *PrivilegeManager {
 }
 
 // DropPrivileges drops root privileges after setup
-func (pm *PrivilegeManager) DropPrivileges(username string) error {
-	if runtime.GOOS == "windows" {
-		// Windows doesn't have the same privilege model
-		return nil
-	}
-	
-	// Check if we're running as root
-	if os.Geteuid() != 0 {
-		return fmt.Errorf("not running as root, cannot drop privileges")
-	}
-	
-	// Look up the target user
-	targetUser, err := user.Lookup(username)
-	if err != nil {
-		return fmt.Errorf("failed to lookup user %s: %w", username, err)
-	}
-	
-	// Parse UID and GID
-	uid, err := strconv.Atoi(targetUser.Uid)
-	if err != nil {
-		return fmt.Errorf("failed to parse UID: %w", err)
-	}
-	
-	gid, err := strconv.Atoi(targetUser.Gid)
-	if err != nil {
-		return fmt.Errorf("failed to parse GID: %w", err)
-	}
-	
-	pm.targetUID = uid
-	pm.targetGID = gid
-	
-	// Set supplementary groups
-	if err := syscall.Setgroups([]int{gid}); err != nil {
-		return fmt.Errorf("failed to set groups: %w", err)
-	}
-	
-	// Set GID
-	if err := syscall.Setgid(gid); err != nil {
-		return fmt.Errorf("failed to set GID: %w", err)
-	}
-	
-	// Set UID (this must be done last)
-	if err := syscall.Setuid(uid); err != nil {
-		return fmt.Errorf("failed to set UID: %w", err)
-	}
-	
-	// Verify privileges were dropped
-	if os.Geteuid() == 0 || os.Getegid() == 0 {
-		return fmt.Errorf("failed to drop root privileges")
-	}
-	
-	return nil
-}
+// Implementation is platform-specific and found in privilege_unix.go and privilege_windows.go
 
 // SetupCapabilities sets Linux capabilities for packet capture
 func (pm *PrivilegeManager) SetupCapabilities() error {
@@ -97,64 +40,9 @@ func (pm *PrivilegeManager) SetupCapabilities() error {
 }
 
 // CheckPrivileges checks if we have necessary privileges
-func (pm *PrivilegeManager) CheckPrivileges() error {
-	switch runtime.GOOS {
-	case "linux", "darwin":
-		if os.Geteuid() != 0 {
-			// Check for capabilities on Linux
-			if runtime.GOOS == "linux" {
-				if err := pm.checkLinuxCapabilities(); err != nil {
-					return fmt.Errorf("insufficient privileges: %w", err)
-				}
-			} else {
-				return fmt.Errorf("requires root privileges")
-			}
-		}
-	case "windows":
-		// Check for administrator
-		if !pm.isWindowsAdmin() {
-			return fmt.Errorf("requires administrator privileges")
-		}
-	default:
-		return fmt.Errorf("unsupported operating system: %s", runtime.GOOS)
-	}
-	
-	return nil
-}
+// Implementation is platform-specific and found in privilege_unix.go and privilege_windows.go
 
-// checkLinuxCapabilities checks for Linux capabilities
-func (pm *PrivilegeManager) checkLinuxCapabilities() error {
-	// Try to execute a capability check
-	// This is a simplified check - real implementation would use libcap
-	cmd := exec.Command("getcap", os.Args[0])
-	output, err := cmd.Output()
-	if err != nil {
-		return fmt.Errorf("no capabilities set (run: sudo setcap cap_net_raw,cap_net_admin+eip %s)", os.Args[0])
-	}
-	
-	outputStr := string(output)
-	for _, cap := range pm.capabilities {
-		if !containsCapability(outputStr, cap) {
-			return fmt.Errorf("missing capability: %s", cap)
-		}
-	}
-	
-	return nil
-}
 
-// isWindowsAdmin checks if running as Windows administrator
-func (pm *PrivilegeManager) isWindowsAdmin() bool {
-	// This is a simplified check
-	// Real implementation would use Windows API
-	_, err := os.Open("\\\\.\\PHYSICALDRIVE0")
-	return err == nil
-}
-
-// containsCapability checks if output contains a capability
-func containsCapability(output, capability string) bool {
-	// Simplified check - real implementation would parse properly
-	return len(output) > 0 // Placeholder
-}
 
 // Sandbox provides process sandboxing
 type Sandbox struct {
@@ -189,45 +77,10 @@ func NewSandbox(workDir string) *Sandbox {
 }
 
 // Enter enters the sandbox
-func (s *Sandbox) Enter() error {
-	if runtime.GOOS == "windows" {
-		// Windows sandboxing would use different mechanisms
-		return nil
-	}
-	
-	// Change working directory
-	if err := os.Chdir(s.workDir); err != nil {
-		return fmt.Errorf("failed to change directory: %w", err)
-	}
-	
-	// Set resource limits
-	if err := s.setResourceLimits(); err != nil {
-		return fmt.Errorf("failed to set resource limits: %w", err)
-	}
-	
-	// On Linux, we could use:
-	// - seccomp for system call filtering
-	// - namespaces for isolation
-	// - cgroups for resource control
-	
-	return nil
-}
+// Implementation is platform-specific and found in privilege_unix.go and privilege_windows.go
 
 // setResourceLimits sets process resource limits
-func (s *Sandbox) setResourceLimits() error {
-	// Set file descriptor limit
-	var rLimit syscall.Rlimit
-	rLimit.Cur = s.resourceLimits["RLIMIT_NOFILE"]
-	rLimit.Max = s.resourceLimits["RLIMIT_NOFILE"]
-	
-	if err := syscall.Setrlimit(syscall.RLIMIT_NOFILE, &rLimit); err != nil {
-		return fmt.Errorf("failed to set RLIMIT_NOFILE: %w", err)
-	}
-	
-	// Additional limits would be set similarly
-	
-	return nil
-}
+// Implementation is platform-specific and found in privilege_unix.go and privilege_windows.go
 
 // SecureExec provides secure command execution
 type SecureExec struct {
@@ -237,54 +90,10 @@ type SecureExec struct {
 }
 
 // NewSecureExec creates a new secure executor
-func NewSecureExec() *SecureExec {
-	return &SecureExec{
-		validator: NewValidator(),
-		allowedCmds: map[string]bool{
-			"ip":      true,
-			"ifconfig": true,
-			"netstat": true,
-			"ss":      true,
-		},
-		environmentVars: []string{
-			"PATH=/usr/bin:/bin",
-			"USER=nobody",
-		},
-	}
-}
+// Implementation is platform-specific and found in privilege_unix.go and privilege_windows.go
 
 // Execute runs a command securely
-func (se *SecureExec) Execute(cmdName string, args ...string) ([]byte, error) {
-	// Validate command is allowed
-	if !se.allowedCmds[cmdName] {
-		return nil, fmt.Errorf("command not allowed: %s", cmdName)
-	}
-	
-	// Validate arguments
-	for _, arg := range args {
-		if containsShellMetachars(arg) {
-			return nil, fmt.Errorf("argument contains shell metacharacters")
-		}
-	}
-	
-	// Create command with clean environment
-	cmd := exec.Command(cmdName, args...)
-	cmd.Env = se.environmentVars
-	
-	// Set security attributes
-	if runtime.GOOS != "windows" {
-		cmd.SysProcAttr = &syscall.SysProcAttr{
-			// Drop supplementary groups
-			Credential: &syscall.Credential{
-				Uid: uint32(os.Getuid()),
-				Gid: uint32(os.Getgid()),
-			},
-		}
-	}
-	
-	// Execute with timeout
-	return cmd.Output()
-}
+// Implementation is platform-specific and found in privilege_unix.go and privilege_windows.go
 
 // SecureDefaults provides secure default configurations
 type SecureDefaults struct {
