+++ pkg/security/privilege_windows.go
@@ -1,0 +1,126 @@
+//go:build windows
+
+package security
+
+import (
+	"fmt"
+	"os"
+	"os/exec"
+)
+
+// DropPrivileges drops root privileges after setup (Windows implementation)
+func (pm *PrivilegeManager) DropPrivileges(username string) error {
+	// Windows doesn't have the same privilege model as Unix
+	// In Windows, we would use different mechanisms like:
+	// - LogonUser API to get user token
+	// - ImpersonateLoggedOnUser to change security context
+	// - CreateProcessAsUser to run with different credentials
+	
+	// For now, this is a no-op since Windows privilege model is different
+	return nil
+}
+
+// CheckPrivileges checks if we have necessary privileges (Windows implementation)
+func (pm *PrivilegeManager) CheckPrivileges() error {
+	// Check for administrator privileges
+	if !pm.isWindowsAdmin() {
+		return fmt.Errorf("requires administrator privileges")
+	}
+	
+	return nil
+}
+
+// isWindowsAdmin checks if running as Windows administrator
+func (pm *PrivilegeManager) isWindowsAdmin() bool {
+	// This is a simplified check for Windows administrator privileges
+	// Real implementation would use Windows API like IsUserAnAdmin() or CheckTokenMembership()
+	// For now, we try to access a system resource that requires admin privileges
+	_, err := os.Open("\\\\.\\PHYSICALDRIVE0")
+	return err == nil
+}
+
+// Enter enters the sandbox (Windows implementation)
+func (s *Sandbox) Enter() error {
+	// Windows sandboxing would use different mechanisms like:
+	// - Job Objects for resource limits
+	// - Restricted tokens for privilege restriction  
+	// - AppContainer for isolation
+	
+	// Change working directory
+	if err := os.Chdir(s.workDir); err != nil {
+		return fmt.Errorf("failed to change directory: %w", err)
+	}
+	
+	// Set resource limits using Windows-specific mechanisms
+	if err := s.setResourceLimits(); err != nil {
+		return fmt.Errorf("failed to set resource limits: %w", err)
+	}
+	
+	return nil
+}
+
+// setResourceLimits sets process resource limits (Windows implementation)
+func (s *Sandbox) setResourceLimits() error {
+	// On Windows, resource limits would be set using:
+	// - Job Objects (SetInformationJobObject)
+	// - Process quotas
+	// - Memory limits via SetProcessWorkingSetSize
+	
+	// For now, this is a placeholder
+	// Real implementation would use Windows API calls
+	
+	return nil
+}
+
+// Execute runs a command securely (Windows implementation)
+func (se *SecureExec) Execute(cmdName string, args ...string) ([]byte, error) {
+	// Validate command is allowed
+	if !se.allowedCmds[cmdName] {
+		return nil, fmt.Errorf("command not allowed: %s", cmdName)
+	}
+	
+	// Validate arguments
+	for _, arg := range args {
+		if containsShellMetachars(arg) {
+			return nil, fmt.Errorf("argument contains shell metacharacters")
+		}
+	}
+	
+	// Create command with clean environment
+	cmd := exec.Command(cmdName, args...)
+	
+	// Use Windows-appropriate environment variables
+	windowsEnv := []string{
+		"PATH=" + os.Getenv("SystemRoot") + "\\System32;" + os.Getenv("SystemRoot"),
+		"COMPUTERNAME=" + os.Getenv("COMPUTERNAME"),
+		"SystemRoot=" + os.Getenv("SystemRoot"),
+	}
+	cmd.Env = windowsEnv
+	
+	// On Windows, we would use different security attributes:
+	// - CreationFlags for process creation options
+	// - ProcessAttributes for security settings
+	// - CreateProcess with restricted tokens
+	
+	// Execute with timeout
+	return cmd.Output()
+}
+
+// NewSecureExec creates a new secure executor (Windows implementation)
+func NewSecureExec() *SecureExec {
+	return &SecureExec{
+		validator: NewValidator(),
+		allowedCmds: map[string]bool{
+			"netstat.exe": true,
+			"ipconfig.exe": true,
+			"ping.exe":    true,
+			"tracert.exe": true,
+			"netsh.exe":   true,
+		},
+		environmentVars: []string{
+			"PATH=" + os.Getenv("SystemRoot") + "\\System32;" + os.Getenv("SystemRoot"),
+			"COMPUTERNAME=" + os.Getenv("COMPUTERNAME"),
+			"SystemRoot=" + os.Getenv("SystemRoot"),
+		},
+	}
+}
